This is the fourth commit, where I have created code that passes my second test and have ran the test to ensure it works.

+++ testing_code/test_driven_code.py
@@ -3,4 +3,6 @@
 
 def find_avg(a,b,c,d,e,f):
     sum = a + b + c + d + e + f
-    return sum / 6
+
+    #The int() function is added in response to the third commit.
+    return int(sum / 6)
